refactor: 결과없음 ui common 아래로 이동
